implement single-responsibility
